requirements.txt
# =============================================================================
# requirements.txt – Concierge Touristique Agent
# =============================================================================
openai>=1.30.0
supabase>=2.4.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0